+++ libs/core/src/lib/adapters/adapter.type.ts
@@ -1,8 +1,16 @@
 // import { Reactions } from './reactions.interface';
+import { SelectorsCache } from '../selectors/memoize-selectors.function';
 import { Selectors } from '../selectors/selectors.interface';
 
 export interface ReactionsWithSelectors<State, S extends Selectors<State>> {
-  [index: string]: ((state: State, event: any, initialState: State) => State) | S;
+  [index: string]:
+    | ((
+        state: State,
+        event: any,
+        initialState: State,
+        selectorsCache: SelectorsCache,
+      ) => State)
+    | S;
 }
 
 type WithSelectors<S> = { selectors?: S };