+++ libs/core/src/lib/adapters/reaction.type.ts
@@ -1,1 +1,8 @@
-export type Reaction<State> = (state: State, event: any, initialState: State) => State;
+import { SelectorsCache } from '../selectors/memoize-selectors.function';
+
+export type Reaction<State> = (
+  state: State,
+  event: any,
+  initialState: State,
+  selectorsCache: SelectorsCache,
+) => State;