+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bugfix(core): undefined is filtered from selections, so TS needed to reflect that.
 - Bugfix(core): Fix buildAdapter type inference
 - Refactor(core)!: Remove deprecated buildSelectors function
+- Feature(core): Pass selectorsCache into state reactions
 
 ## 0.59.0 2022-11-09